% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_landcover2015_lookup.R
\name{can_landcover2015_lookup}
\alias{can_landcover2015_lookup}
\title{Look-up Canada Landcover 2015 values and RGB colours.}
\usage{
can_landcover2015_lookup()
}
\value{
A data.frame with Canada Landcover 2015 look-up information.
}
\description{
See Canada Landcover 2015 raster values, associated landcover descriptions, and standard RGB colour codes.
}
\examples{
\dontrun{
can_landcover2015_lookup()
}
}
