% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{degree_to_percent}
\alias{degree_to_percent}
\title{Convert slope degrees to percent}
\usage{
degree_to_percent(degrees)
}
\arguments{
\item{degrees}{Vector of slope values in degrees. For example, the output of calling \code{raster::terrain(opt = 'slope', unit = 'degrees')}.}
}
\value{
Vector of slope values in percent.
}
\description{
Convert slope degrees to percent slope.
}
\examples{
\dontrun{
library(raster)
library(sf)

# create reference raster
r <- raster(xmn = -1198869, xmx = -1131697, ymn = 901665.8, ymx = 1004528, res = 250, crs = 3978)
elev <- elev_grab(reference_grid = r, output_directory = NULL)
plot(elev)

# get slope
slope_degrees <- raster::terrain(elev, opt = 'slope', unit = 'degrees')
slope_percent <- degree_to_percent(slope_degrees) # now slope is represented as percent
}
}
