% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Variables_grab.R
\name{ecozone_grab}
\alias{ecozone_grab}
\title{Grab ecozones}
\usage{
ecozone_grab(reference_grid)
}
\arguments{
\item{reference_grid}{A reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object.}
}
\value{
RasterLayer with the name 'Ecozone'.
}
\description{
Extracts and rasterizes a subset of the Terrestrial Ecozones of Canada to a user-defined reference grid.
}
\details{
Extracts and rasterizes Terrestrial Ecozones of Canada to align with the input reference_grid. The original ecozones data are in polygon format, and are downloaded to a tempfile using this function.

See original metadata to interpret raster values at \url{https://open.canada.ca/data/en/dataset/7ad7ea01-eb23-4824-bccc-66adb7c5bdf8}
}
\examples{
\dontrun{
library(raster)
library(sf)

# create reference raster
r <- raster(xmn = -1198869, xmx = -1131697, ymn = 901665.8, ymx = 1004528, res = 250, crs = 3978)
ecozones <- ecozone_grab(reference_grid = r)
plot(ecozones)
}
}
\references{
Open Data Canada: \url{https://open.canada.ca/data/en/dataset/7ad7ea01-eb23-4824-bccc-66adb7c5bdf8}
}
