% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecozone_lookup.R
\name{ecozone_lookup}
\alias{ecozone_lookup}
\title{Look-up Canada Ecozone values and RGB colours.}
\usage{
ecozone_lookup()
}
\value{
A data.frame with Canada Ecozone look-up information.
}
\description{
See Canada Ecozone raster values, associated names, and RGB colour codes.
}
\examples{
\dontrun{
ecozone_lookup()
}
}
