% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Variables_grab.R
\name{elev_grab}
\alias{elev_grab}
\title{Grab elevation}
\usage{
elev_grab(reference_grid, output_directory = NULL, filename_prefix = NULL)
}
\arguments{
\item{reference_grid}{A reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object.}

\item{output_directory}{Output directory where your elevation grid will be exported. If NULL, the raster will not be exported to the local device.}

\item{filename_prefix}{String that will be appended to the start of the exported raster file name.}
}
\value{
RasterLayer with the name 'Elevation'.
}
\description{
A user-defined reference grid is used to define the NTS grids necessary to pull elevation from the CDEM layers provided by the Government of Canada.
}
\details{
The purpose of this function is to generate a common and rapid elevation layer that is sampled and masked to the reference grid.
The pre-packaged GeoTiff datasets used to generate the output grid are based on the NTS grids at the 1:250 000 scale.
}
\note{
Requires an internet connection
}
\examples{
\dontrun{
library(raster)
library(sf)

# create reference raster
r <- raster(xmn = -1198869, xmx = -1131697, ymn = 901665.8, ymx = 1004528, res = 250, crs = 3978)
elev <- elev_grab(reference_grid = r, output_directory = NULL)
plot(elev)
}
}
\references{
Canadian Digital Elevation Model: \url{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}
}
