% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5Land_GEE_grab.R
\name{ERA5Land_GEE_grab}
\alias{ERA5Land_GEE_grab}
\title{Grab ERA5-Land hourly rasters from Google Earth Engine}
\usage{
ERA5Land_GEE_grab(
  variable = NULL,
  start_date,
  end_date,
  extent,
  buff_width = 0,
  match_crs = FALSE,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{variable}{A string or vector of ERA5 Land hourly raster band names. E.g. \code{c('dewpoint_temperature_2m','temperature_2m')}. Run \code{ERA5Land_VariableList()} to see all accepted options.}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{extent}{An sp, sf, or Raster* object used to define the extraction extent.}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in extent, thereby increasing the extraction extent. Defaults to 0.}

\item{match_crs}{Logical. Should the extracted rasters be projected to the CRS of the reference extent object? Defaults to \code{FALSE.} If \code{FALSE}, the raster is returned in the native projection and resolution (WGS 84, 0.1 arc degrees).}

\item{output_directory}{Optional. Output directory where NetCDFs will be exported. If \code{NULL}, the NetCDFs will not be exported to the local device.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported raster file name. Root names of files are based on variable names.}
}
\value{
A named list of RasterBricks
}
\description{
Extract ERA5-Land hourly rasters using Google Earth Engine to a user-defined date range and spatial extent (via package rgee).
}
\details{
This function extracts ERA5-Land hourly variables to a user-defined date range and spatial extent using the rgee package.

A Google Earth Engine account is required to use this function. For more information, see \url{https://earthengine.google.com/}.
Google drive must also be accessible by Google Earth Engine for downloading purposes. Run \code{ee_initialize(drive = TRUE)} before running this function.
}
\examples{
\dontrun{

library(rgee)
library(raster)
library(sf)
library(lubridate)
library(lutz)
library(humidity)

ee_Initialize(drive = TRUE) # initialize rgee

# define an extent polygon
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\%
     st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

# use ERA5Land_VariableList() to see band options
ERA5Land_VariableList()

temperature_brick <- ERA5Land_GEE_grab(variable = 'temperature_2m',
                                       start_date = '2019-05-20',
                                       end_date = '2019-05-25',
                                       extent = e, # sp or sf or Raster*
                                       buff_width = 0, # meters
                                       match_crs = TRUE,
                                       output_directory = NULL,
                                       filename_prefix = NULL)

plot(temperature_brick[[1]][[1:4]])
}
}
\references{
Google Earth Engine, ERA5-Land hourly: \url{https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_LAND_HOURLY#description}

rgee: \url{https://github.com/r-spatial/rgee}
}
