% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5Land_table_GEE_grab.R
\name{ERA5Land_table_GEE_grab}
\alias{ERA5Land_table_GEE_grab}
\title{Grab ERA5-Land hourly tables from Google Earth Engine}
\usage{
ERA5Land_table_GEE_grab(
  start_date,
  end_date,
  extent,
  buff_width = 0,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{extent}{An sp, sf, or Raster* object used to define the extraction extent.}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in extent, thereby increasing the extraction extent. Defaults to 0.}

\item{output_directory}{Optional. Output directory where a CSV will be exported. If \code{NULL}, the CSV will not be exported to the local device.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported CSV.}
}
\value{
A tibble
}
\description{
Extract ERA5-Land hourly tables using Google Earth Engine to a user-defined date range and spatial extent (via package rgee).
}
\details{
This function extracts ERA5-Land hourly variables (in table format) to a user-defined date range and spatial extent using the rgee package.
In contrast, \code{ERA5Land_GEE_grab()} will extract the hourly rasters.
The benefit of using this \code{ERA5Land_table_GEE_grab()} is improved speed.

This function only extracts a selected subset of weather metrics, including dewpoint_temperature_2m, temperature_2m, snow_cover, total_precipitation_hourly,
surface_solar_radiation_downwards_hourly, u_component_of_wind_10m, v_component_of_wind_10m, and surface_pressure.
Relative humidity and windspeed are automatically calculated from these inputs.
The values in the resulting table represent the MEDIAN value over the extraction extent per hourly timestep.

A Google Earth Engine account is required to use this function. For more information, see \url{https://earthengine.google.com/}.
Google drive must also be accessible by Google Earth Engine for downloading purposes. Run \code{ee_initialize(drive = TRUE)} before running this function.
}
\references{
Google Earth Engine, ERA5-Land hourly: \url{https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_LAND_HOURLY#description}

rgee: \url{https://github.com/r-spatial/rgee}
}
