% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5Land_GEE_grab.R
\name{ERA5Land_VariableList}
\alias{ERA5Land_VariableList}
\title{ERA5-Land hourly Variable List}
\usage{
ERA5Land_VariableList()
}
\value{
data.frame of ERA5 Land variables
}
\description{
See a list of ERA5-Land hourly variables that can be pulled from Google Earth Engine.
}
\details{
Run this function to see a list of ERA5 Land hourly variables that can be pulled using the function \code{ERA5Land_GEE_grab()}.
A vector of the items under the BandName column can be used for the parameter variable in \code{ERA5Land_GEE_grab()}.
Units for each variable are also included.

For more information, see \url{https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_LAND_HOURLY#description}.
}
\examples{
\dontrun{
ERA5Land_VariableList()
}
}
\references{
Google Earth Engine, ERA5-Land hourly: \url{https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_LAND_HOURLY#description}
}
