% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Variables_grab.R
\name{fbp_fuel_grab}
\alias{fbp_fuel_grab}
\title{Grab 2014 FBP fuels grid}
\usage{
fbp_fuel_grab(
  reference_grid,
  reference_poly,
  match_crs = TRUE,
  buff_width = 0,
  fbp_grid = NULL
)
}
\arguments{
\item{reference_grid}{A reference raster used to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object. See details.}

\item{reference_poly}{A reference polygon that provides a projection and extent to extract forest attributes for. It can be an sf or sp polygon object, or the location of a polygon on your local device. See details.}

\item{match_crs}{Logical. Should the extracted forest attribute grid be projected to the CRS of the reference object? Defaults to TRUE. If FALSE, the raster is returned in the native projection, Canada_Lambert_Conformal_Conic.}

\item{buff_width}{Numeric. Width of buffer (in meters) that should be applied to the \code{reference_poly} before extraction. Defaults to 0. Ignored if \code{reference_poly} is missing.}

\item{fbp_grid}{A FBP fuels raster object or the location of the raster on your local device. If \code{NULL}, the FBP fuels 2014 raster will be downloaded to a tempfile. See details for recommendations.}
}
\value{
A RasterLayer with the name 'FBP_fuels_2014'.
}
\description{
Extracts a spatial subset of the 2014 FBP fuels map to a user defined extent.
}
\details{
The purpose of this function is to extract a spatial subset of the 2014 FBP fuels map.

Only one of \code{reference_grid} or \code{reference_poly} must be defined to set the extraction extent. If both are defined, an error will occur.
If \code{reference_grid} is used to define the extraction extent and \code{match_CRS = TRUE}, then the returned raster will align with the reference_grid.
If \code{reference_grid} is used to define the extraction extent and \code{match_CRS = FALSE}, then the returned raster will be in the native grid and projection, Canada_Lambert_Conformal_Conic.

It is recommended to define the \code{fbp_grid} if this function is being called multiple times, in order to reduce download time.
See metadata to interpret raster values at \url{https://cwfis.cfs.nrcan.gc.ca/downloads/fuels/current/}
}
\examples{
\dontrun{
library(raster)
library(sf)

# Define a polygon extent
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\% st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert
fbp_fuel <- fbp_fuel_grab(reference_poly = e, match_crs = TRUE, buff_width = 0, fbp_grid = NULL)
plot(fbp_fuel)
}
}
\references{
Open Data Canada: \url{https://cwfis.cfs.nrcan.gc.ca/downloads/fuels/current/}
}
