% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBP_Fuels2014_lookup.R
\name{FBP_Fuels2014_lookup}
\alias{FBP_Fuels2014_lookup}
\title{Look-up FBP Fuels 2014 Values and RGB colours.}
\usage{
FBP_Fuels2014_lookup()
}
\value{
A data.frame with FBP Fuels 2014 look-up information.
}
\description{
See FBP Fuels 2014 raster values, associated fuel types, and standard RGB colour codes.
}
\examples{
\dontrun{
FBP_Fuels2014_lookup()
}
}
