% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Variables_grab.R
\name{foliage_biomass_grab}
\alias{foliage_biomass_grab}
\title{Grab foliage biomass}
\usage{
foliage_biomass_grab(
  reference_grid,
  reference_poly,
  match_crs = TRUE,
  buff_width = 0,
  foliage_grid = NULL
)
}
\arguments{
\item{reference_grid}{A reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object. See details.}

\item{reference_poly}{A reference polygon that provides a projection and extent to extract forest attributes for. It can be an sf or sp polygon object, or the location of a polygon on your local device. See details.}

\item{match_crs}{Logical. Should the extracted forest attribute grid be projected to the CRS of the reference object? Defaults to TRUE. If FALSE, the raster is returned in the native projection, SR-ORG:8787.}

\item{buff_width}{Numeric. Width of buffer (in meters) that should be applied to the \code{reference_poly} before extraction. Defaults to 0. Ignored if \code{reference_poly} is missing.}

\item{foliage_grid}{A foliage biomass raster object or the location of the raster on your local device. If \code{NULL}, the foliage biomass raster will be downloaded to a tempfile. See details for recommendations.}
}
\value{
RasterLayer with the name 'Foliage_biomass'.
}
\description{
Extract a spatial subset of the foliage biomass map from the 2011 Canada's Forest Attributes to a user defined extent.
}
\details{
The purpose of this function is to extract a spatial subset of the Canada Forest Attributes foliage biomass map.

Only one of \code{reference_grid} or \code{reference_poly} must be defined to set the extraction extent. If both are defined, an error will occur.
If \code{reference_grid} is used to define the extraction extent and \code{match_CRS = TRUE}, then the returned raster will align with the \code{reference_grid.}
If \code{reference_grid} is used to define the extractiion extent and \code{match_CRS = FALSE}, then the returned raster will be in the native grid and projection, SR-ORG:8787.

It is recommended to define the \code{foliage_grid} if this function is being called multiple times, in order to reduce download time.
Values in the returned raster represent t/ha.
}
\examples{
\dontrun{
library(raster)
library(sf)

# define polygon extent
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\% st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert
r <- foliage_biomass_grab(reference_poly = e, match_crs = TRUE, buff_width = 0, foliage_grid = NULL)
plot(r)
}
}
\references{
2011 Canada's Forest Attributes: \url{https://open.canada.ca/data/en/dataset/ec9e2659-1c29-4ddb-87a2-6aced147a990}
}
