% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWI_ERA5_grab.R
\name{FWI_ERA5_grab}
\alias{FWI_ERA5_grab}
\title{Grab FWI metrics}
\usage{
FWI_ERA5_grab(
  variable = NULL,
  overwinter_DC = TRUE,
  start_date,
  end_date,
  extent,
  buff_width = 0,
  match_crs = FALSE,
  input_directory = NULL,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{variable}{A string or vector of FWI metrics. E.g. \code{c('FFMC','DMC','DC','ISI','BUI','FWI')}.}

\item{overwinter_DC}{Logical. Should FWI metrics using overwintered DC be extracted? Default is \code{TRUE}. If \code{FALSE}, FWI metrics using default DC startup is extracted (see details).}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01". Must have the same year as start_date.}

\item{extent}{An sp or sf object used to define the extraction extent.}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in extent, thereby increasing the extraction extent. Defaults to 0.}

\item{match_crs}{Logical. Should the extracted rasters be projected to the CRS of the reference extent object? Defaults to \code{FALSE}. If \code{FALSE}, the raster is returned in the native projection and resolution (WGS 84, 0.25 arc degrees).}

\item{input_directory}{Directory to where FWI NetCDFs are saved on the local device. If \code{NULL}, FWI NetCDFs will be downloaded to tempfiles. See details for recommendations.}

\item{output_directory}{Optional. Output directory where NetCDFs will be exported. If \code{NULL}, the NetCDFs will not be exported to the local device.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported raster file name.}
}
\value{
List of named RasterBricks
}
\description{
Extract daily rasters for Fire Weather Indices calculated using ERA5 Reanalysis data to a user-defined date range and spatial extent.
}
\details{
This function extracts daily FWI rasters (from \url{https://zenodo.org/record/3626193}) to a user-defined date range and spatial extent.
Currently, this function can only temporally subset to dates within the same year.

It is recommended to download NetCDFs for the variables and years of interest from \url{https://zenodo.org/record/3626193}, then define the \code{input_directory} to reduce download times.
If the \code{input_directory} is defined, this function will search for the proper NetCDF using both year and variable name within the \code{input_directory.} The user should not change the default name of files downloaded from the Zenodo page.
}
\examples{
\dontrun{
library(raster)
library(sf)
library(lubridate)
library(dplyr)
library(ncdf4)
library(curl)

# Create extent polygon
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\%
     st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

FWI <- FWI_ERA5_grab(variable = 'FWI',# string or vector of variables, e.g. c('FWI','FFMC')
                    overwinter_DC = TRUE,
                    start_date = '2018-05-20',
                    end_date='2018-05-30', # must be same year as start date
                    extent = e, # sp or sf object with extent and crs info
                    buff_width = 0,
                    match_crs = TRUE, # match crs of input extent object?
                    input_directory = NULL, # folder where ncdfs are saved
                    output_directory = NULL)

plot(FWI[[1]][[1:4]])
}
}
\references{
Global Fire Weather Indices: \url{https://zenodo.org/record/3626193}
}
