% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_canada_timezones}
\alias{get_canada_timezones}
\title{Import Canada timezones}
\usage{
get_canada_timezones(location = NULL, crs = 4326)
}
\arguments{
\item{location}{A spatial object of class sf or sp. If provided, the intersecting timezones will be returned. If this parameter is not provided, timezone polygons for all of Canada are returned.}

\item{crs}{Calls \code{st_transform()} to project the timezone polygons into a desired coordinate system. Valid inputs to the \code{crs} argument of \code{st_transform()} are allowed. Defaults to WGS84 (EPSG 4326).}
}
\value{
sf polygons object
}
\description{
Import Canada timezones into R
}
\details{
Downloads Canada timezones to a temporary file and imports the timezone polygons into the R environment.
Timezones are taken from \url{https://github.com/evansiroky/timezone-boundary-builder}

Additional attributes beyond the ones provided from \url{https://github.com/evansiroky/timezone-boundary-builder}
are included in order to facilitate compability with the R packages lutz and lubridate.
Often, local standard times (LST) for hotspots and weather observations are needed for wildfire applications. Both lutz and lubridate
only recognize official Olson names (see \code{OlsonNames())}) as timezones (e.g. "America/Edmonton"). These timezones force daylight savings,
even when standard times are desired.

Accordingly, the following attributes are available from the sf object imported using \code{canada_timezones()}:

\code{tzid}: The timezone ID as an Olson name. For example, "America/Edmonton".
These names were pulled from \url{https://github.com/evansiroky/timezone-boundary-builder}

\code{LST_offset}: The local standard time (LST) UTC offset that corresponds with the tzid.
UTC offsets are positive east of UTC and negative west of UTC.

\code{LDT_offset}: The local daylight time (LDT) UTC offset that corresponds with the tzid.
If the timezone does not experience daylight savings, then this field is the same as the \code{LST_offset}.
UTC offsets are positive east of UTC and negative west of UTC.

\code{LST_tzid}: An Olson name timezone ID that corresponds with the LST_offset.
For example, an LST_offset of "-07:00" would correspond to the timezone "Etc/GMT+7".
The purpose of this is to allow for compatiability with the R packages lutz and lubridate, which only recognizes official Olson name timezones.
Accordingly, offsets that do not have a corresponding Olson name are NA. This is the case for America/St_Johns (LST_offset = -03:30).

\code{LDT_tzid}: An Olson name timezone ID that corresponds with the LDT_offset.
For example, an LDT_offset of "-07:00" would correspond to the timezone "Etc/GMT+7".
The purpose of this is to allow for compatiability with the R packages lutz and lubridate, which only recognizes official Olson name timezones.
Accordingly, offsets that do not have a corresponding Olson name are NA. This is the case for America/St_Johns (LDT_offset = -02:30).

If an area experiences daylight savings, then the user can instead use \code{LST_tzid} in the lubridate package to "force"
the equivalent local standard time.
}
