% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_CFS_ID}
\alias{get_CFS_ID}
\title{Get CFS REF ID}
\usage{
get_CFS_ID(NBAC)
}
\arguments{
\item{NBAC}{An sp or sf NBAC polygon of one fire event.}
}
\value{
sf points
}
\description{
Get CFS REF ID of NBAC fire perimeters.
}
\details{
In order to extract the CFS REF ID of NBAC fire perimeters, this function downloads the CNFDB fire point data to a tempfile,
then subsets the CNFDB points temporally and spatially to parameters defined within the input \code{NBAC}.
The CFS REF ID of the CNFDB point with the largest area is then returned.

Users should be cautious when using output from this function; it is simply a guess at what the CFS REF ID for a certain NBAC fire should be.
Other sources should be checked to validate the CFS REF ID output from this function.
}
