% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOES_FRP_grab.R
\name{GOES_FRP_grab}
\alias{GOES_FRP_grab}
\title{Grab GOES-16 and GOES-17 FRP rasters from Google Earth Engine}
\usage{
GOES_FRP_grab(
  satellite,
  extent,
  buff_width = 0,
  start_date,
  end_date,
  match_crs = FALSE,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{satellite}{Specify 16 for GOES-16, or 17 for GOES-17.}

\item{extent}{An sp, sf, or Raster* object used to define the extraction extent.}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in extent, thereby increasing the extraction extent. Defaults to 0.}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{match_crs}{Logical. Should the extracted rasters be projected to the CRS of the reference extent object? Defaults to \code{FALSE}. If \code{FALSE}, the raster is returned in the native projection and resolution (GOES-R or GOES-S ABI fixed grid projection, res = 2000 m).}

\item{output_directory}{Optional. Output directory where a NetCDF will be exported. If \code{NULL}, the NetCDF will not be exported to the local device.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported raster file name.}
}
\value{
A RasterBrick
}
\description{
Extract GOES-16 or GOES-17 FDCF Series ABI Level 2 FRP rasters using Google Earth Engine from the R environment (via package rgee).
}
\details{
This function extracts GOES-16 or GOES-17 10-minute rasters to a user-defined date interval and spatial extent.
A Google Earth Engine account is required to use this function. For more information, see \url{https://earthengine.google.com/}.
Google drive must also be accessible by Google Earth Engine for downloading purposes. Run \code{ee_initialize(drive = TRUE)} before running this function.

A separate request is made for each unique month of data requested. This is to limit the amount of data transferred with any one request. All requests are then merged and returned as a single RasterBrick.
Data is updated in near real-time (~ 30 minute lag).

For more information, see
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description}
and
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}
}
\examples{
\dontrun{
library(raster)
library(sf)
library(sp)
library(lubridate)
library(rgee)
library(lutz)
library(rgdal)

ee_Initialize(drive = TRUE)

# create extent polygon
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\% st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

# import GOES-16 FRP rasters as RasterBrick in R environment
GOES16_FRP <- GOES_FRP_grab(satellite = 16, # either 16 for GOES-16, or 17 for GOES-17
                            extent = e,
                            buff_width = 2000, # meters
                            start_date = '2019-05-20',
                            end_date = '2019-05-21',
                            match_crs = TRUE,
                            output_directory = NULL)

# plot 10-minute rasters
plot(GOES16_FRP[[1:4]])

# to visualize better, sum all layers together
GOES16_FRP_sum <- stackApply(GOES16_FRP, indices = rep(1,nlayers(GOES16_FRP)), sum)
plot(GOES16_FRP_sum)
}
}
\references{
Google Earth Engine, GOES-16 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description}

Google Earth Engine, GOES-17 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}

rgee: \url{https://github.com/r-spatial/rgee}
}
