% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOES_FRP_table_grab.R
\name{GOES_FRP_table_grab}
\alias{GOES_FRP_table_grab}
\title{Grab GOES-16 and GOES-17 FRP tables from Google Earth Engine}
\usage{
GOES_FRP_table_grab(
  reference_poly,
  buff_width = 0,
  start_date,
  end_date,
  interval = "Hourly"
)
}
\arguments{
\item{reference_poly}{A sf or sp polygon that defines the spatial extent that FRP values are summed within.}

\item{buff_width}{Width (in meters) to buffer the \code{reference_poly}. Defaults to 0.}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{interval}{Time interval to sum FRP values over. Choices are \code{"Daily"}, \code{"Hourly"}, or \code{"10min"}. Defaults to \code{"Hourly"}. See details for definiton of \code{"Daily"} aggregation.}
}
\value{
A tibble
}
\description{
Extract GOES-16 or GOES-17 FDCF Series ABI Level 2 FRP tables using Google Earth Engine from the R environment (via package rgee).
}
\details{
This function extracts GOES-16 and GOES-17 FRP values within a user-defined date interval and spatial polygon.
The returned table values represent the sum of FRP values within the specified time interval in the spatial polygon provided.

If \code{interval = "Daily"}, the interval is defined as a 24 hr period starting at 9AM local time and ending at 8:59AM on the subsequent day.

A Google Earth Engine account is required to use this function. For more information, see \url{https://earthengine.google.com/}.

Data is updated in near real-time (~ 30 minute lag). For more information, see
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description} and
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}
}
\examples{
\dontrun{
library(raster)
library(sf)
library(lubridate)
library(rgee)
library(lutz)
library(tidyr)

ee_Initialize()

# Create extent polygon
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\% st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

GOES_tbl <- GOES_FRP_table_grab(reference_poly = e,
                               buff_width = 0,
                               start_date = '2019-05-20',
                               end_date = '2019-05-25',
                               interval = 'Hourly')

GOES_tbl
}
}
\references{
Google Earth Engine, GOES-16 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description}

Google Earth Engine, GOES-17 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}

rgee: \url{https://github.com/r-spatial/rgee}
}
