% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOES_GEE_grab.R
\name{GOES_GEE_grab}
\alias{GOES_GEE_grab}
\title{Grab GOES point data from Google Earth Engine}
\usage{
GOES_GEE_grab(
  satellite,
  maskvals = c(10, 11, 12, 13, 14, 15, 30, 31, 32, 33, 34, 35),
  extent,
  buff_width = 0,
  start_date,
  end_date,
  return_sf = TRUE,
  match_crs = FALSE
)
}
\arguments{
\item{satellite}{Specify 16 for GOES-16, or 17 for GOES-17.}

\item{maskvals}{Vector of maskvals to retain in extraction. Default is \code{c(10,11,12,13,14,15,30,31,32,33,34,35)}. See details.}

\item{extent}{An sp, sf, or Raster* object used to define the extraction extent.}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in \code{extent}, thereby increasing the extraction extent. Defaults to 0.}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{return_sf}{Logical. Should the points be returned as an sf object? If \code{FALSE}, returns a tibble with lat/lon information. Defaults to TRUE.}

\item{match_crs}{Logical. Should the extracted points be projected to the CRS of the reference \code{extent} object? Only applicable if \code{return_sf = TRUE}. Defaults to \code{FALSE}. If \code{FALSE}, sf points are returned in WGS84.}
}
\value{
A tibble or sf points object.
}
\description{
Extract GOES-16 or GOES-17 FDCF Series ABI Level 2 data as points using Google Earth Engine from the R environment (via package rgee).
}
\details{
This function extracts GOES-16 or GOES-17 FDCF Series ABI Level 2 rasters as points to a user-defined date interval and spatial extent.
A Google Earth Engine account is required to use this function. For more information, see \url{https://earthengine.google.com/}.
Google drive must also be accessible by Google Earth Engine for downloading purposes. Run \code{ee_initialize(drive = TRUE)} before running this function.

A single data transfer call to GEE is made if less than 30 days of data are requested. If > 30 days are requested, multiple calls to GEE are made.
Data is updated in near real-time (~ 30 minute lag).

\code{maskvals} is a vector that denotes which GOES raster mask values to retain. For more information on mask values, see page 530 of the GOES Users' Guide: \url{https://www.goes-r.gov/products/docs/PUG-L2+-vol5.pdf}.
To retain ALL mask values, use \code{maskvals = 'all'}

For more information, see
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description}
and
\url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}
}
\examples{
\dontrun{
library(sf)
library(lubridate)
library(rgee)
library(tibble)
library(data.table)
library(dplyr)
library(tmap)

ee_Initialize(drive = TRUE)

# create extent polygon
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\% st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

# import GOES-16 FDCF product into the R environment as sf points, in WGS84:
goes_points <- GOES_GEE_grab(satellite = 16, # '16' or '17'
                             maskvals = c(10,11,12,13,14,15,30,31,32,33,34,35), #maskvals to keep
                             extent = e, # sf, sp or Raster* object to grab extent from
                             buff_width = 8000, # will buffer the extent of the extent above
                             start_date = '2019-05-19',
                             end_date = '2019-05-22',
                             return_sf = TRUE, # if TRUE, return sf
                             match_crs = FALSE) #if return_sf = TRUE, match crs of extent?

goes_points # returns 2,614 features

# visualize
tm_shape(e) + tm_borders() +
  tm_shape(goes_points) + tm_dots() +
  tm_layout(frame = F)

# import GOES-16 FDCF product into the R environment as a tibble:
goes_tibble <- GOES_GEE_grab(satellite = 16,
                             maskvals = c(10,11,12,13,14,15,30,31,32,33,34,35),
                             extent = e,
                             buff_width = 8000,
                             start_date = '2019-05-19',
                             end_date = '2019-05-22',
                             return_sf = FALSE)

goes_tibble # returns 2,614 features. latlon info given.

# keep only highest confidence mask values:
goes_tibble_conf <- GOES_GEE_grab(satellite = 16,
                             maskvals = c(10,11,30,31),
                             extent = e,
                             buff_width = 8000,
                             start_date = '2019-05-19',
                             end_date = '2019-05-22',
                             return_sf = FALSE)

goes_tibble_conf # returns 743 features (subset only to highest confidence). latlon info given.
}
}
\references{
Google Earth Engine, GOES-16 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_16_FDCF#description}

Google Earth Engine, GOES-17 FDCF: \url{https://developers.google.com/earth-engine/datasets/catalog/NOAA_GOES_17_FDCF#description}

rgee: \url{https://github.com/r-spatial/rgee}
}
