% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_Variables_grab.R
\name{landcover2015_grab}
\alias{landcover2015_grab}
\title{Landcover grab}
\usage{
landcover2015_grab(reference_grid, landcover_grid = NULL)
}
\arguments{
\item{reference_grid}{A reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object.}

\item{landcover_grid}{A landcover raster object or the location of a landcover raster on your local device. If NULL, the 2015 Land Cover of Canada will be downloaded to a tempfile (not recommended, see Details).}
}
\value{
RasterLayer with the name 'Landcover'.
}
\description{
Extract a spatial subset of the 2015 Land Cover of Canada to a user-defined reference grid.
}
\details{
The purpose of this function is to extract a spatial subset of the 2015 Land Cover of Canada that aligns with your input reference grid.
It is highly recommended to download the 2015 Land Cover of Canada to your local device and use the landcover_grid argument to access it, especially if this function is being run multiple times.
The 2015 Land Cover grid is 1.9 GB and time consuming to download through this function.

To interpret values in the returned raster, refer to \url{https://open.canada.ca/data/en/dataset/4e615eae-b90c-420b-adee-2ca35896caf6}.
}
\examples{
\dontrun{
library(raster)
library(sf)

# create reference raster
r <- raster(xmn = -1198869, xmx = -1131697, ymn = 901665.8, ymx = 1004528, res = 250, crs = 3978)
lc <- landcover2015_grab(reference_grid = r)
plot(lc)
}
}
\references{
2015 Land Cover of Canada: \url{https://open.canada.ca/data/en/dataset/4e615eae-b90c-420b-adee-2ca35896caf6}
}
