% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIIRS_MODIS_hotspot_grab.R
\name{MODIS_hotspots_grab}
\alias{MODIS_hotspots_grab}
\title{Grab MODIS hotspots}
\usage{
MODIS_hotspots_grab(
  reference_poly,
  match_crs = TRUE,
  start_date,
  end_date,
  buff_width = 0
)
}
\arguments{
\item{reference_poly}{An sp or sf polygon, defining the extent that hotspots are extracted for. The location of a polygon file is also accepted.}

\item{match_crs}{Logical. Should the extracted hotspots be projected to the CRS of the \code{reference_poly}? Defaults to \code{TRUE}. If \code{FALSE}, the hotspots are returned in WGS 84.}

\item{start_date}{Date object or character. Start date of data extraction, using UTC. E.g."2018-05-01".}

\item{end_date}{Date object or character. End date of data extraction, using UTC. E.g."2018-05-07".}

\item{buff_width}{Numeric. Width (in meters) to buffer the object defined in extent, thereby increasing the extraction extent. Defaults to 0.}
}
\value{
sf points
}
\description{
Import and/or download archived MODIS Collection 6 hotspots from the FIRMS database within a user-defined date range and spatial extent.
}
\details{
This function extracts the archived MODIS Collection 6 hotspots from \url{https://firms.modaps.eosdis.nasa.gov/download/}.
The Country Yearly Summary CSVs are used for download, which means there is a considerable time lag until hotspots are available using this function (> 1 year lag).

Currently, this function does not support download of NRT MODIS Collection 6 hotspots.
}
\examples{
\dontrun{
library(sf)
library(raster)
library(lubridate)
library(dplyr)
library(tibble)

# Define a polygon extent
e <- extent(-1198869,-1131697,901665.8,1004528) \%>\%
     st_bbox() \%>\% st_as_sfc() \%>\% st_sf()
st_crs(e) <- 3978 # assign CRS as EPSG 3978, Canada Atlas Lambert

modis <- MODIS_hotspots_grab(reference_poly = e, start_date = '2019-05-20', end_date = '2019-06-10')
plot(modis[,'acq_date'])
}
}
\references{
FIRMS: \url{https://firms.modaps.eosdis.nasa.gov/download/}
}
