% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{round_choose}
\alias{round_choose}
\title{Round numbers up/down to any interval}
\usage{
round_choose(x, y)
}
\arguments{
\item{x}{Numeric. Value to be rounded up or down to specified interval.}

\item{y}{Numeric. Interval used to round \code{x}. Positive \code{y} values denote rounding up, whereas negative values denote rounding down.}
}
\value{
Numeric; rounded value of \code{x}.
}
\description{
Round numbers up/down to user defined interval.
}
\details{
For example, using \code{y = 5} will round \code{x} up to the nearest multiple of 5, whereas using \code{y = -5} will round \code{x} down to the nearest multiple of 5.
}
\examples{
\dontrun{
round_choose(27.8, 2) # round UP to the nearest 2
round_choose(27.8, -2) # round DOWN to the nearest 2
round_choose(38497, 10) # round UP to the nearest 10
round_choose(38497, -10) # round DOWN to the nearest 10
}
}
