% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{searchNBAC_by_alias}
\alias{searchNBAC_by_alias}
\title{Search NBAC by alias}
\usage{
searchNBAC_by_alias(nbac_file, alias)
}
\arguments{
\item{nbac_file}{Location of an NBAC file or NBAC polygon object. Download NBAC from \url{https://cwfis.cfs.nrcan.gc.ca/datamart/metadata/nbac}.}

\item{alias}{A string that can contain the fire alias or CFS REF ID. This string is searched for matches within the NBAC Comments field.}
}
\value{
sf polygon
}
\description{
Searches the NBAC comments field using a user-defined alias string to return matching polygons.
}
\examples{
\dontrun{
library(sf)
library(dplyr)

# download 2019 NBAC to tempfile
nbac_temp <- tempfile(fileext = '.zip')
download.file(destfile = nbac_temp,
              url = 'https://cwfis.cfs.nrcan.gc.ca/downloads/nbac/nbac_2019_r9_20200703.zip')

nbac_files <- unzip(nbac_temp, list = T)
nbac <- unzip(zipfile = nbac_temp, files = nbac_files$Name[1:11], exdir = gsub('.zip','',nbac_temp))
nbac <- st_read(dsn = nbac[7])

# search for the 2019 McMillan wildfire, CFS ID SWF-049-2019
fire <- searchNBAC_by_alias(nbac_file = nbac, alias = 'SWF-049-2019')

plot(fire[,1]) # view the McMillan fire
}
}
\references{
NBAC: \url{https://cwfis.cfs.nrcan.gc.ca/datamart/metadata/nbac}
}
