% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{simplify_NBAC}
\alias{simplify_NBAC}
\title{Simplify NBAC}
\usage{
simplify_NBAC(original_NBAC)
}
\arguments{
\item{original_NBAC}{sf polygon of an NBAC fire perimeter. Probably the output from \code{searchNBAC_by_alias()}}
}
\value{
sf polygon
}
\description{
Simplify NBAC perimeters by dissolving the constituent polygons can calling \code{rmapshaper::ms_simplify()}
}
\details{
This function simplifies NBAC perimeters to make them faster to work with for the Data Cubes for Canadian Wildfires project.

Some NBAC field attributes are retained. Included retained attributes are YEAR, NFIREID, AGENCY, SDATE, EDATE.
Some attributes are added, including SEPOCH (Start EPOCH), EEPOCH (End EPOCH), NBAC_HA (final total size in ha).

If an alias is available in the comments, this is also extracted to a new field called NAME (e.g. 'SWF-049-2019' has the NAME 'McMillan Complex Fire).
The \code{original_NBAC} is also simplified using default settings of \code{rmapshaper::ms_simplify()}, then the resulting polygons are dissolved together.
}
\examples{
\dontrun{
library(sf)
library(stringr)
library(rmapshaper)
library(dplyr)

# download 2019 NBAC to tempfile
nbac_temp <- tempfile(fileext = '.zip')
download.file(destfile = nbac_temp,
              url = 'https://cwfis.cfs.nrcan.gc.ca/downloads/nbac/nbac_2019_r9_20200703.zip')

nbac_files <- unzip(nbac_temp, list = T)
nbac <- unzip(zipfile = nbac_temp, files = nbac_files$Name[1:11], exdir = gsub('.zip','',nbac_temp))
nbac <- st_read(dsn = nbac[7])

# search for the 2019 McMillan wildfire, CFS ID SWF-049-2019
fire <- searchNBAC_by_alias(nbac_file = nbac, alias = 'SWF-049-2019')

plot(fire[,1]) # view the McMillan fire

# simplify the geometry and dissolve into one polygon
NBAC_simplify <- simplify_NBAC(fire)
NBAC_simplify # simplified polygon, easier to work with
plot(NBAC_simplify[,1]) # notice simplified polygon looks pretty much the same as original
}
}
\references{
NBAC: \url{https://cwfis.cfs.nrcan.gc.ca/datamart/metadata/nbac}
}
