% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{split_1hr}
\alias{split_1hr}
\title{Split intervals into 1-hour intervals}
\usage{
split_1hr(time_range)
}
\arguments{
\item{time_range}{A lubridate interval object.}
}
\value{
A vector of lubridate intervals.
}
\description{
Split lubridate interval objects into 1-hour intervals.
}
\examples{
\dontrun{
library(lubridate)
int <- interval(start = '2019-05-10', end = '2019-05-11')
int_split_1hr <- split_1hr(int)
int_split_1hr
}
}
