% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{split_24hrs}
\alias{split_24hrs}
\title{Split intervals into 24-hour intervals}
\usage{
split_24hrs(time_range)
}
\arguments{
\item{time_range}{A lubridate interval object.}
}
\value{
A vector of lubridate intervals.
}
\description{
Split lubridate interval objects into 24-hour intervals.
}
\examples{
\dontrun{
library(lubridate)
int <- interval(start = '2019-05-10', end = '2019-05-20')
int_split_24hr <- split_24hrs(int)
int_split_24hr
}
}
