% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fire_vars.R
\name{summarize_daily_fire_vars}
\alias{summarize_daily_fire_vars}
\title{Summarize daily fire variables}
\usage{
summarize_daily_fire_vars(
  interim_stack,
  NBAC = NULL,
  VIIRS_hotspots = NULL,
  MODIS_hotspots = NULL,
  GOES_table = NULL,
  FWI_list = NULL,
  variable_stack = NULL,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{interim_stack}{Output of calling \code{IDW_interpolate_hotspots(interval = 'Daily', cumulative = F)}. A RasterBrick or RasterStack containing non-cumulative estimates of daily area burned, where 1 = burned during the daily interval and 0 = not burned during the daily interval.}

\item{NBAC}{sf polygon of the final NBAC fire perimeter. Likely the output of calling \code{simplify_NBAC()}.}

\item{VIIRS_hotspots}{sf points of VIIRS hotspots. The output of calling \code{VIIRS_hotspots_grab()}.}

\item{MODIS_hotspots}{sf points of MODIS hotspots. The output of calling \code{MODIS_hotspots_grab()}.}

\item{GOES_table}{GOES FRP table. The output of calling \code{GOES_FRP_table(interval = 'Hourly')}.}

\item{FWI_list}{List of FWI RasterBricks. The output of calling \code{FWI_ERA5_grab()}.}

\item{variable_stack}{Named RasterBrick or RasterStack of topography and landcover rasters. See details.}

\item{output_directory}{Optional. Output directory where a CSV should be saved.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported CSV. Ignored if \code{output_directory = NULL}.}
}
\value{
A tibble
}
\description{
Summarize daily fire variables, such as hotspot counts, FRP, and FWI metrics on a daily timeframe based on estimated area burned.
}
\details{
In brief, this function summarizes hotspot and FWI data on a daily timeframe based on estimated area of burn (as determined by IDW interpolation of VIIRS hotspots).
Note that "daily" summaries refer to a 24 hr period from 9AM on a given day to 8:59AM on the subsequent day (a "day" does not begin at midnight).

This function requires very specific inputs, which are derived from running other functions created for the Data Cubes for Canadian Wildfires project.
Only specific names of items in the variable_stack are recognized, including "Ecozone", "FBP_fuels_2014", "Landcover", "Elevation", "Aspect", "Slope", "Canopy_closure", "Stand_height", "Broadleaf_percentage", "Needleleaf_percentage", "Branch_biomass", "Foliage_biomass".
Using the provided functions to extract static rasters will ensure layers are named correctly.

To see an example of this function in use, please refer to the example extraction workflow script.
}
