% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fire_vars.R
\name{summarize_hourly_fire_vars}
\alias{summarize_hourly_fire_vars}
\title{Summarize hourly fire variables}
\usage{
summarize_hourly_fire_vars(
  interim_stack,
  NBAC = NULL,
  VIIRS_hotspots = NULL,
  MODIS_hotspots = NULL,
  GOES16_hotspots = NULL,
  GOES17_hotspots = NULL,
  ERA5land_weather_list = NULL,
  output_directory = NULL,
  filename_prefix = NULL
)
}
\arguments{
\item{interim_stack}{Output of calling \code{IDW_interpolate_hotspots(interval = 'Daily', cumulative = F)}. A RasterBrick or RasterStack containing non-cumulative estimates of daily area burned, where 1 = burned during the daily interval and 0 = not burned during the daily interval.}

\item{NBAC}{sf polygon of the final NBAC fire perimeter. Likely the output of calling \code{simplify_NBAC()}.}

\item{VIIRS_hotspots}{sf points of VIIRS hotspots. The output of calling \code{VIIRS_hotspots_grab()}.}

\item{MODIS_hotspots}{sf points of MODIS hotspots. The output of calling \code{MODIS_hotspots_grab()}.}

\item{GOES16_hotspots}{sf points of GOES-16 hotspots. The output of calling \code{GOES_GEE_grab(satellite = 16)}.}

\item{GOES17_hotspots}{sf points of GOES-17 hotspots. The output of calling \code{GOES_GEE_grab(satellite = 17)}.}

\item{ERA5land_weather_list}{List of ERA5 Land hourly RasterBricks. The output of calling \code{ERA5Land_GEE_grab()}.}

\item{output_directory}{Optional. Output directory where a CSV should be saved.}

\item{filename_prefix}{Optional. String that will be appended to the start of the exported CSV. Ignored if \code{output_directory = NULL}.}
}
\value{
A tibble
}
\description{
Summarize hourly fire variables, such as hotspot counts, FRP, and weather on an hourly timeframe based on estimated area burned.
}
\details{
This function requires very specific inputs, which are derived from running other functions created for the Data Cubes for Canadian Wildfires project.
In brief, this function summarizes hotspot and weather data on an hourly timeframe based on estimated area of burn (as determined by IDW interpolation of VIIRS hotspots).
Since there is no hourly estimate of area burned, all estimated burn pixels for the entire day that the hourly interval falls within are used for summarizing.

To see an example of this function in use, please refer to the example extraction workflow script.
}
